brute CRUD

+++ src/main/java/com/adira/contact/service/UserServiceImpl.java
@@ -1,8 +1,13 @@
 package com.adira.contact.service;
 
+import com.adira.contact.common.Utils;
+import com.adira.contact.exception.CustomNotFoundException;
 import com.adira.contact.pojo.User;
 import com.adira.contact.repository.UserRepository;
 
+import jakarta.transaction.Transactional;
+
+import org.apache.coyote.BadRequestException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -29,24 +34,45 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User updateUser(Long id, User user) {
-        Optional<User> existingUserOptional = userRepository.findById(id);
+    @Transactional
+    public String updateUser(Long id, User user) throws BadRequestException {
+        User existingUser = userRepository.findById(id)
+                .orElseThrow(() -> new CustomNotFoundException("User with ID " + id + " not found"));
 
-        if (existingUserOptional.isPresent()) {
-            User existingUser = existingUserOptional.get();
-            // Update fields as needed
+        if (user.getEmail() != null && !user.getEmail().equals(existingUser.getEmail())) {
+            if (doesUserExistByEmail(user.getEmail())) {
+                throw new CustomNotFoundException("Email " + user.getEmail() + " already exists!");
+            }
             existingUser.setEmail(user.getEmail());
-            existingUser.setPassword(user.getPassword());
-            existingUser.setAdmin(user.isAdmin());
-
-            return userRepository.save(existingUser);
         }
 
-        return null; // Handle the case where the user with the specified ID is not found
+        if (user.getPassword() != null) {
+            Utils utils = new Utils();
+            if (!utils.validatePassword(user.getPassword())) {
+                throw new BadRequestException(
+                        "Password must contain at least 8 characters, including one uppercase letter, one lowercase letter, one digit, and one special character");
+            } else {
+                existingUser.setPassword(user.getPassword());
+            }
+        }
+
+        userRepository.save(existingUser);
+
+        return "User with ID " + id + " updated successfully";
     }
 
     @Override
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+        } else {
+            throw new CustomNotFoundException("User with ID " + id + " not found");
+        }
+
+    }
+
+    @Override
+    public boolean doesUserExistByEmail(String email) {
+        return userRepository.existsByEmail(email);
     }
 }
